🔧 dish table view cell을 재사용하기 위해 stack view로 새로 구현 (#11)
