+++ SideDishWithRx/SideDishWithRx/View/DishTableViewCell.swift
@@ -15,35 +15,11 @@ final class DishTableViewCell: UITableViewCell {
         let imageView = UIImageView()
         return imageView
     }()
-    private var title = UILabel()
-    private let dishDescription = UILabel()
-    private let priceStackView: UIStackView = {
-        var stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.alignment = .leading
-        stackView.spacing = 4
-        return stackView
-    }()
-    private let nPrice = UILabel()
-    private let sPrice = UILabel()
-    private let badgeStackView: UIStackView = {
-        var stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.alignment = .leading
-        stackView.spacing = 4
-        return stackView
-    }()
-    private let dishInformationStackView: UIStackView = {
-        var stackView = UIStackView()
-        stackView.axis = .vertical
-        stackView.distribution = .fillEqually
-        stackView.alignment = .leading
-        return stackView
-    }()
+
+    private let dishInformationStackView = DishInformationStackView()
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
-        configureDishInformationStackView()
         self.addSubviews()
         self.autoLayout()
     }
@@ -60,8 +36,7 @@ final class DishTableViewCell: UITableViewCell {
     }
     
     override func prepareForReuse() {
-        priceStackView.subviews.forEach { $0.removeFromSuperview() }
-        badgeStackView.subviews.forEach { $0.removeFromSuperview() }
+        dishInformationStackView.prepareForReuseAsCell()
     }
     
     private func addSubviews() {
@@ -69,58 +44,8 @@ final class DishTableViewCell: UITableViewCell {
         contentView.addSubview(dishInformationStackView)
     }
     
-    private func configureDishInformationStackView() {
-        dishInformationStackView.addArrangedSubview(title)
-        dishInformationStackView.addArrangedSubview(dishDescription)
-        dishInformationStackView.addArrangedSubview(priceStackView)
-    }
-    
-    private func configurePriceStackView(nPriceValue: String?, sPriceValue: String) {
-        sPrice.text = sPriceValue
-        priceStackView.addArrangedSubview(sPrice)
-
-        
-        if let nPriceValue = nPriceValue {
-            let attributedString = changeStrikeThrough(text: nPriceValue)
-            nPrice.attributedText = attributedString
-            nPrice.textColor = .gray
-            priceStackView.addArrangedSubview(nPrice)
-        }
-    }
-    
-    private func changeStrikeThrough(text: String) -> NSAttributedString {
-        let attributeString = NSMutableAttributedString(string: text)
-        attributeString.addAttribute(.strikethroughStyle, value: 1, range: NSMakeRange(0, attributeString.length))
-        return attributeString
-    }
-    
-    private func configureBadgeStackView(badge: [String?]?) {
-        guard let badge = badge, !badge.compactMap({ $0 }).isEmpty else {
-            return
-        }
-
-        badge.forEach { badge in
-            if badge == "이벤트특가" {
-                let badgeLabel = BadgeLabel(frame: CGRect(x: 0, y: 0, width: 0, height: 0), backgroundColor: .systemIndigo, topInset: 5, bottomInset: 5, leadingInset: 7, trailingInset: 7)
-                badgeLabel.text = badge
-                badgeStackView.addArrangedSubview(badgeLabel)
-            } else {
-                let badgeLabel = BadgeLabel(frame: CGRect(x: 0, y: 0, width: 0, height: 0), backgroundColor: .systemYellow, topInset: 5, bottomInset: 5, leadingInset: 7, trailingInset: 7)
-                badgeLabel.text = badge
-                badgeStackView.addArrangedSubview(badgeLabel)
-            }
-        }
-        
-        dishInformationStackView.addArrangedSubview(badgeStackView)
-    }
-    
     func configureCell(title: String, description: String, nprice: String?, sPrice: String, badge: [String?]?) {
-        self.title.text = title
-        self.dishDescription.text = description
-        
-        configurePriceStackView(nPriceValue: nprice, sPriceValue: sPrice)
-        configureBadgeStackView(badge: badge)
-        
+        dishInformationStackView.configureCell(title: title, description: description, nprice: nprice, sPrice: sPrice, badge: badge)
     }
 }
